requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.9.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.1
djangorestframework==3.13.1
finnhub-python==2.4.14
frozenlist==1.3.1
idna==3.3
multidict==6.0.2
psycopg2-binary==2.9.3
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
yarl==1.8.1
